pip_requirements.txt
psycopg2-binary
SQLAlchemy
python-binance
